% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_bolt4jr.R
\name{setup_bolt4jr}
\alias{setup_bolt4jr}
\title{Set up the Conda environment for bolt4jr}
\usage{
setup_bolt4jr()
}
\description{
This function initializes the Conda environment required for the \code{bolt4jr} package.
If no Conda binary is found, it installs Miniconda. If the required Conda environment
(\code{bolt4jr}) is not found, it creates the environment and installs the necessary dependencies.
}
\details{
The function ensures that:
\itemize{
\item A Conda binary is available.
\item A Conda environment named \code{bolt4jr} exists.
\item The \code{neo4j} Python package is installed in the \code{bolt4jr} environment.
}

Call this function manually before using any functionality that relies on Python.
}
\examples{
setup_bolt4jr()

}
